fix(input): input require ngModel (#UIM-361)
