+++ packages/mosaic/form-field/public-api.ts
@@ -1,7 +1,6 @@
 export * from './form-field.module';
 export * from './form-field';
 export * from './form-field-control';
-export * from './form-field-number-control';
 export * from './form-field-errors';
 export * from './hint';
 export * from './suffix';
