+++ packages/mosaic/input/input-value-accessor.ts
@@ -1,5 +1,4 @@
 import { InjectionToken } from '@angular/core';
 
 
-export const MC_INPUT_VALUE_ACCESSOR =
-    new InjectionToken<{ value: any }>('MC_INPUT_VALUE_ACCESSOR');
+export const MC_INPUT_VALUE_ACCESSOR = new InjectionToken<{ value: any }>('MC_INPUT_VALUE_ACCESSOR');
