+++ packages/mosaic/input/public-api.ts
@@ -1,4 +1,4 @@
 export * from './input.module';
 export * from './input';
-export * from './stepperUtils';
+export * from './input-number';
 export * from './input-value-accessor';
